+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-**Version 0.6.1 (unreleased)**
+**Version 0.7.0 (unreleased)**
  - Minor compression improvement on interlaced images
  - Performance optimizations
+ - [SEMVER_MINOR] Move default Options into a Default impl
 
 **Version 0.6.0**
  - Fix issue where output directory would not be created if it did not exist